+++ tests/0200_cli/0201_cli_version.py
@@ -1,0 +1,16 @@
+from click.testing import CliRunner
+
+from py_css_styleguide.cli.entrypoint import cli_frontend
+
+
+def test_version_ping(caplog):
+    """
+    Just ping version command, should always be successful without any logs.
+    """
+    runner = CliRunner()
+
+    result = runner.invoke(cli_frontend, ["version"])
+
+    assert result.exit_code == 0
+
+    assert caplog.record_tuples == []